add: Added for theme creation process
