+++ laravel/app/Http/Controllers/ThemaController.php
@@ -2,6 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\ThemaRequest;
+use App\Models\Thema;
+use App\UseCase\Thema\ShowThemaListUseCase;
+use App\UseCase\Thema\StoreThemaUseCase;
 use Illuminate\Http\Request;
 
 class ThemaController extends Controller
@@ -9,11 +13,16 @@ class ThemaController extends Controller
   /**
    * Display a listing of the resource.
    *
-   * @return \Illuminate\Http\Response
+   * @return
    */
-  public function index()
+  public function index(Request $request, ShowThemaListUseCase $useCase)
   {
+    // ユーザログイン機能を作るまでひとまず「１」を入れる
+    $user_id = 1;
 
+    return $useCase
+      ? response()->json($useCase($user_id), 201)
+      : response()->json([], 500);
   }
 
   /**
@@ -23,7 +32,7 @@ class ThemaController extends Controller
    */
   public function create()
   {
-      //
+
   }
 
   /**
@@ -32,9 +41,13 @@ class ThemaController extends Controller
    * @param  \Illuminate\Http\Request  $request
    * @return \Illuminate\Http\Response
    */
-  public function store(Request $request)
+  public function store(ThemaRequest $request, StoreThemaUseCase $useCase)
   {
-      //
+    // $thema->fill($request->all())->save();
+
+    return $useCase
+      ? response()->json($useCase($request), 201)
+      : response()->json([], 500);
   }
 
   /**
