+++ laravel/app/Models/Thema.php
@@ -9,6 +9,11 @@ class Thema extends Model
 {
     use HasFactory;
 
+    protected $fillable = [
+        'user_id',
+        'thema',
+    ];
+
     public function innerWord()
     {
         return $this->hasMany(InnerWord::class);
